1.更新部分代码

+++ lib/DevApp/src/main/java/dev/utils/common/ArrayUtils.java
@@ -1899,7 +1899,7 @@ public final class ArrayUtils {
                 JCLogUtils.eTag(TAG, e, "get");
             }
         }
-        return (double) -1;
+        return -1D;
     }
 
     /**
